Split Bracken estimation into a separate piepline
