+++ q2_moshpit/__init__.py
@@ -6,11 +6,11 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from .kraken2 import classification, database
+from .kraken2 import bracken, classification, database
 from .metabat2 import metabat2
 
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-__all__ = ['metabat2', 'classification', 'database']
+__all__ = ['metabat2', 'bracken', 'classification', 'database']
